chore: apply machine-applicable clippy fixes across the workspace (#19)
`cargo clippy --fix` over the tree (needless borrows, index loops → iterators,
redundant casts/closures, derivable/Default impls, collapsed let-chains,
excessive-precision literals, unused imports) — 70 files, behavior-preserving.
Plus the `crdt` loro deprecation: `get_or_create_container(k, LoroText::new())`
→ `ensure_mergeable_text(k)`.

Does NOT clear the advisory clippy job fully — correctness-level lints
(approx_constant, erasing_op, bad_bit_mask in DSP) and pre-existing broken
--all-targets example builds remain; tracked separately for review.

Verified: cargo check --workspace --exclude vox-discover is clean.

+++ libs/nice-plug/nice-plug-dioxus/src/editor.rs
@@ -181,7 +181,7 @@ impl raw_window_handle::HasWindowHandle for RwhAdapter {
         let raw = match self.0 {
             ParentWindowHandle::X11Window(window) => {
                 let handle = raw_window_handle::XcbWindowHandle::new(
-                    NonZeroU32::new(window as u32).expect("X11 window ID should not be 0"),
+                    NonZeroU32::new(window).expect("X11 window ID should not be 0"),
                 );
                 RawWindowHandle::Xcb(handle)
             }
